requirements.txt
numpy==1.26.4
Flask==3.0.0
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12